Made conections.
Created reference to LayerGenerator and ArtefactGen in the dropdown script. Separated JSON classes from dropdown script.
